@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -11,21 +10,21 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20150103014259) do
+ActiveRecord::Schema.define(version: 2015_01_03_014259) do
 
   create_table "lemonade_types", force: :cascade do |t|
-    t.string   "label"
-    t.string   "color"
-    t.decimal  "price"
+    t.string "label"
+    t.string "color"
+    t.decimal "price"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
 
   create_table "purchases", force: :cascade do |t|
-    t.integer  "quantity"
-    t.integer  "lemonade_type_id"
-    t.datetime "created_at",       null: false
-    t.datetime "updated_at",       null: false
+    t.integer "quantity"
+    t.integer "lemonade_type_id"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
 end
